Rename input-field block to form-input
