+++ docs/reference-guides/core-blocks.md
@@ -286,6 +286,15 @@ A form. ([Source](https://github.com/WordPress/gutenberg/tree/trunk/packages/blo
 -	**Supports:** anchor, color (background, gradients, link, text), spacing (margin, padding), typography (fontSize, lineHeight), ~~className~~
 -	**Attributes:** action, formId, method
 
+## Input field
+
+The basic building block for forms. ([Source](https://github.com/WordPress/gutenberg/tree/trunk/packages/block-library/src/form-input))
+
+-	**Name:** core/form-input
+-	**Category:** common
+-	**Supports:** 
+-	**Attributes:** inlineLabel, label, name, required, type
+
 ## Classic
 
 Use the classic WordPress editor. ([Source](https://github.com/WordPress/gutenberg/tree/trunk/packages/block-library/src/freeform))
@@ -349,15 +358,6 @@ Insert an image to make a visual statement. ([Source](https://github.com/WordPre
 -	**Category:** media
 -	**Supports:** anchor, color (~~background~~, ~~text~~), filter (duotone)
 -	**Attributes:** align, alt, aspectRatio, caption, height, href, id, lightbox, linkClass, linkDestination, linkTarget, rel, scale, sizeSlug, title, url, width
-
-## Input field
-
-The basic building block for forms. ([Source](https://github.com/WordPress/gutenberg/tree/trunk/packages/block-library/src/input-field))
-
--	**Name:** core/input-field
--	**Category:** common
--	**Supports:** 
--	**Attributes:** inlineLabel, label, name, required, type
 
 ## Latest Comments
