+++ lib/blocks.php
@@ -22,9 +22,9 @@ function gutenberg_reregister_core_block_types() {
 				'column',
 				'columns',
 				'details',
+				'form-input',
 				'group',
 				'html',
-				'input-field',
 				'list',
 				'list-item',
 				'media-text',
