+++ packages/block-library/src/form-input/block.json
@@ -1,7 +1,7 @@
 {
 	"$schema": "https://schemas.wp.org/trunk/block.json",
 	"apiVersion": 2,
-	"name": "core/input-field",
+	"name": "core/form-input",
 	"title": "Input field",
 	"category": "common",
 	"parent": [ "core/form" ],
@@ -29,5 +29,5 @@
 			"default": false
 		}
 	},
-	"style": [ "wp-block-input-field", "wp-block-buttons", "wp-block-button" ]
+	"style": [ "wp-block-form-input", "wp-block-buttons", "wp-block-button" ]
 }
