+++ packages/block-library/src/index.js
@@ -49,13 +49,13 @@ import * as details from './details';
 import * as embed from './embed';
 import * as file from './file';
 import * as form from './form';
+import * as formInput from './form-input';
 import * as gallery from './gallery';
 import * as group from './group';
 import * as heading from './heading';
 import * as homeLink from './home-link';
 import * as html from './html';
 import * as image from './image';
-import * as inputField from './input-field';
 import * as latestComments from './latest-comments';
 import * as latestPosts from './latest-posts';
 import * as list from './list';
@@ -232,7 +232,7 @@ const getAllBlocks = () => {
 	];
 	if ( window?.__experimentalEnableFormBlocks ) {
 		blocks.push( form );
-		blocks.push( inputField );
+		blocks.push( formInput );
 	}
 
 	// When in a WordPress context, conditionally
