+++ packages/block-library/src/style.scss
@@ -14,11 +14,11 @@
 @import "./details/style.scss";
 @import "./embed/style.scss";
 @import "./file/style.scss";
+@import "./form-input/style.scss";
 @import "./gallery/style.scss";
 @import "./group/style.scss";
 @import "./heading/style.scss";
 @import "./image/style.scss";
-@import "./input-field/style.scss";
 @import "./latest-comments/style.scss";
 @import "./latest-posts/style.scss";
 @import "./list/style.scss";
