+++ test/integration/fixtures/blocks/core__form-input.serialized.html
@@ -1,0 +1,3 @@
+<!-- wp:form-input {"label":"Name","required":true} -->
+<label class="wp-block-form-input-label"><div class="wp-block-form-input-label__content">Name</div><input class="wp-block-form-input" type="text" name="Name" required="" aria-required="true"></label>
+<!-- /wp:form-input -->
