+++ test/integration/fixtures/blocks/core__form.html
@@ -1,21 +1,21 @@
 <!-- wp:form -->
-<form><!-- wp:input-field {"label":"Name","required":true} -->
-<label class="wp-block-input-field-label"><div class="wp-block-input-field-label__content">Name</div><input class="wp-block-input-field" type="text" name="Name" required aria-required="true"/></label>
-<!-- /wp:input-field -->
+<form><!-- wp:form-input {"label":"Name","required":true} -->
+<label class="wp-block-form-input-label"><div class="wp-block-form-input-label__content">Name</div><input class="wp-block-form-input" type="text" name="Name" required aria-required="true"/></label>
+<!-- /wp:form-input -->
 
-<!-- wp:input-field {"type":"email","label":"Email","required":true} -->
-<label class="wp-block-input-field-label"><div class="wp-block-input-field-label__content">Email</div><input class="wp-block-input-field" type="email" name="Email" required aria-required="true"/></label>
-<!-- /wp:input-field -->
+<!-- wp:form-input {"type":"email","label":"Email","required":true} -->
+<label class="wp-block-form-input-label"><div class="wp-block-form-input-label__content">Email</div><input class="wp-block-form-input" type="email" name="Email" required aria-required="true"/></label>
+<!-- /wp:form-input -->
 
-<!-- wp:input-field {"type":"url","label":"Website"} -->
-<label class="wp-block-input-field-label"><div class="wp-block-input-field-label__content">Website</div><input class="wp-block-input-field" type="url" name="Website" aria-required="false"/></label>
-<!-- /wp:input-field -->
+<!-- wp:form-input {"type":"url","label":"Website"} -->
+<label class="wp-block-form-input-label"><div class="wp-block-form-input-label__content">Website</div><input class="wp-block-form-input" type="url" name="Website" aria-required="false"/></label>
+<!-- /wp:form-input -->
 
-<!-- wp:input-field {"type":"textarea","label":"Comment","required":true} -->
-<label class="wp-block-input-field-label"><div class="wp-block-input-field-label__content">Comment</div><textarea class="wp-block-input-field" name="Comment" required aria-required="true"></textarea></label>
-<!-- /wp:input-field -->
+<!-- wp:form-input {"type":"textarea","label":"Comment","required":true} -->
+<label class="wp-block-form-input-label"><div class="wp-block-form-input-label__content">Comment</div><textarea class="wp-block-form-input" name="Comment" required aria-required="true"></textarea></label>
+<!-- /wp:form-input -->
 
-<!-- wp:input-field {"type":"submit","label":"Submit"} -->
+<!-- wp:form-input {"type":"submit","label":"Submit"} -->
 <div class="wp-block-buttons"><div class="wp-block-button"><button class="wp-block-button__link wp-element-button">Submit</button></div></div>
-<!-- /wp:input-field --></form>
+<!-- /wp:form-input --></form>
 <!-- /wp:form -->
