--- test/integration/fixtures/blocks/core__input-field.html
@@ -1,3 +1,0 @@
-<!-- wp:input-field {"label":"Name","required":true} -->
-<label class="wp-block-input-field-label"><div class="wp-block-input-field-label__content">Name</div><input class="wp-block-input-field" type="text" name="Name" required aria-required="true"/></label>
-<!-- /wp:input-field -->
