--- test/integration/fixtures/blocks/core__input-field.json
@@ -1,12 +1,0 @@
-[
-	{
-		"name": "core/missing",
-		"isValid": true,
-		"attributes": {
-			"originalName": "core/input-field",
-			"originalUndelimitedContent": "<label class=\"wp-block-input-field-label\"><div class=\"wp-block-input-field-label__content\">Name</div><input class=\"wp-block-input-field\" type=\"text\" name=\"Name\" required aria-required=\"true\"/></label>",
-			"originalContent": "<!-- wp:input-field {\"label\":\"Name\",\"required\":true} -->\n<label class=\"wp-block-input-field-label\"><div class=\"wp-block-input-field-label__content\">Name</div><input class=\"wp-block-input-field\" type=\"text\" name=\"Name\" required=\"\" aria-required=\"true\"></label>\n<!-- /wp:input-field -->"
-		},
-		"innerBlocks": []
-	}
-]
