--- test/integration/fixtures/blocks/core__input-field.parsed.json
@@ -1,14 +1,0 @@
-[
-	{
-		"blockName": "core/input-field",
-		"attrs": {
-			"label": "Name",
-			"required": true
-		},
-		"innerBlocks": [],
-		"innerHTML": "\n<label class=\"wp-block-input-field-label\"><div class=\"wp-block-input-field-label__content\">Name</div><input class=\"wp-block-input-field\" type=\"text\" name=\"Name\" required aria-required=\"true\"/></label>\n",
-		"innerContent": [
-			"\n<label class=\"wp-block-input-field-label\"><div class=\"wp-block-input-field-label__content\">Name</div><input class=\"wp-block-input-field\" type=\"text\" name=\"Name\" required aria-required=\"true\"/></label>\n"
-		]
-	}
-]
